Cleanup
* Add header generation
* Move files to separate directories
