+++ include/smtp/mimemultipart.h
@@ -21,7 +21,7 @@
 
 #include "mimepart.h"
 
-#include "smtpexports.h"
+#include "smtp_export.h"
 
 class SMTP_EXPORT MimeMultiPart : public MimePart
 {
